requirements.txt
aerich @ git+https://github.com/tortoise/aerich.git@c60c1610f0dfdb4a8a38ecaf1af49a846d5de2c3
aiosqlite==0.17.0
anyio==3.4.0
asgiref==3.4.1
asyncpg==0.25.0
click==8.0.3
ddlparse==1.10.0
dictdiffer==0.9.0
fastapi==0.70.1
h11==0.12.0
httptools==0.3.0
idna==3.3
iso8601==0.1.16
pydantic==1.8.2
pyparsing==3.0.6
pypika-tortoise==0.1.1
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
sniffio==1.2.0
starlette==0.16.0
tomlkit==0.7.2
tortoise-orm==0.17.8
typing-extensions==4.0.1
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
